Infraguardian/backend/app/requirements.txt
langgraph>=0.2.36
openai>=1.40.0
langchain>=0.2.12
langchain-openai>=0.1.17
pydantic>=2.7.0
requests>=2.32.3
python-dotenv>=1.0.1